+++ esp-radio/src/ble/btdm.rs
@@ -83,7 +83,7 @@ extern "C" fn notify_host_recv(data: *mut u8, len: u16) -> i32 {
 
     super::dump_packet_info(data);
 
-    crate::ble::controller::asynch::hci_read_data_available();
+    crate::ble::controller::hci_read_data_available();
 
     0
 }