+++ esp-radio/src/ble/npl.rs
@@ -1377,7 +1377,7 @@ unsafe extern "C" fn ble_hs_hci_rx_evt(cmd: *const u8, arg: *const c_void) -> i3
         r_ble_hci_trans_buf_free(cmd);
     }
 
-    crate::ble::controller::asynch::hci_read_data_available();
+    crate::ble::controller::hci_read_data_available();
 
     0
 }
@@ -1406,7 +1406,7 @@ unsafe extern "C" fn ble_hs_rx_data(om: *const OsMbuf, arg: *const c_void) -> i3
         r_os_mbuf_free_chain(om as *mut _);
     }
 
-    crate::ble::controller::asynch::hci_read_data_available();
+    crate::ble::controller::hci_read_data_available();
 
     0
 }