+++ esp-radio/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `dtim_period` parameter for `PowerSaveMode` (#4040)
 - `WifiConfig`, `CountryInfo` and `OperatingClass` (#4121)
 - Configuration options for `BleController` (#4223, #4254, #4259)
+- BLE controller: Added support for `embedded-io 0.7` (#4280)
 
 ### Changed